+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1776 Error out when index SQ8H run in CPU mode
 -   \#1929 Skip MySQL meta schema field width check
 -   \#1946 Fix load index file CPU2GPU fail during searching
+-   \#1955 Switch create_index operation to background once client break connection
 -   \#1997 Index file missed after compact
 -   \#2073 Fix CheckDBConfigBackendUrl error message
 -   \#2076 CheckMetricConfigAddress error message